Refactor code out of rblcheck into libdnscheck
This is to enable re-use and -- I hope -- make it easier to test.  No
tests yet, though.
