+++ libdnscheck/Cargo.toml
@@ -7,3 +7,9 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+thiserror = "~1.0"
+dbus = "0.9.1"
+dbus-tree = "0.9.0"
+libc = "0.2.86"
+
+libdbusdnscheck = { path = "../libdbusdnscheck" }
