+++ rblcheck/Cargo.toml
@@ -20,4 +20,4 @@ dbus = "0.9.1"
 dbus-tree = "0.9.0"
 libc = "0.2.86"
 
-libdbusdnscheck = { path = "../libdbusdnscheck" }
+libdnscheck = { path = "../libdnscheck" }
